bt-127: * swapped names of modals and relocated them in one folder
