--- mobile/src/bundles/common/components/active-modal/styles.ts
@@ -1,14 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-const styles = StyleSheet.create({
-    backdrop: {
-        backgroundColor: 'rgba(0,0,0,0.5)',
-    },
-    container: {
-        width: '80%',
-        gap: 20,
-        backgroundColor: '#fff',
-    },
-});
-
-export { styles };
