+++ mobile/src/bundles/common/components/components.tsx
@@ -1,10 +1,10 @@
-export { ActiveModal } from './active-modal/active-modal';
 export { Button } from './button/button';
 export { FormField } from './form-field/form-field';
 export { Input } from './input/input';
 export { Link } from './link/link';
 export { Loader } from './loader/loader';
-export { Modal } from './modal/modal';
+export { ActiveModal } from './modals/active-modal';
+export { Modal } from './modals/modal';
 export { Overlay } from './overlay/overlay';
 export { Text } from './text/text';
 export {
